move ValueToJSON out of archive
