+++ src/EncryptedDevice/CMakeLists.txt
@@ -8,5 +8,12 @@ target_sources(EncryptedDevice
         EncryptedDevice.hpp
 )
 
+# TODO: ValueToJSON is only used by CLI and C API, it does not belong here.
+target_sources(EncryptedDevice
+    PRIVATE
+        ValueToJSON.cpp
+        ValueToJSON.hpp
+)
+
 target_include_directories(EncryptedDevice INTERFACE "${CMAKE_CURRENT_LIST_DIR}/..")
 target_link_libraries(EncryptedDevice TrustedPeripheral)
